clean src, and config
