+++ benchmark/run.ts
@@ -1,15 +1,16 @@
 import Benchmark from "benchmark";
-import {benchmarkCases} from "../tests/testCases";
+
+import { benchmarkCases } from "../tests/testCases";
 import lightClassNames from "../";
 
-const suite = new Benchmark.Suite('classnames official benchmark');
+const suite = new Benchmark.Suite('light-classnames benchmark');
 
-suite.on("cycle", function(event: Benchmark.Event) {
+suite.on("cycle", (event: Benchmark.Event) => {
     console.log(String(event.target));
 });
 
 benchmarkCases.forEach((testCase) => {
-    suite.add(testCase.name, function() {
+    suite.add(testCase.name, () => {
         lightClassNames(testCase.value);
     });
 });
